Split dev constants
